MICRSE-2981: cst_signer: Add support for signing images that do not contain SPL
This patch fixes the sign procedure such that an itb image can be given
as argument.
It looks like the HAB search string is not robust enough. For an itb that
contains a kernel image it is likely to find patterns that resembles with IVT.
In this case the condition is that the ivt should be in the end of the image.

Signed-off-by: Radu Bulie <radu-andrei.bulie@nxp.com>